requirements.txt
asgiref==3.3.4
autopep8==1.5.6
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
defusedxml==0.7.1
Django==3.2
django-braces==1.14.0
django-cors-headers==3.7.0
django-environ==0.4.5
django-filter==2.4.0
django-oauth-toolkit==1.5.0
djangorestframework==3.12.4
drf-social-oauth2==1.0.9
idna==2.10
itypes==1.2.0
Jinja2==2.11.3
jwcrypto==0.8
MarkupSafe==1.1.1
oauthlib==3.1.0
Pillow==8.2.0
pycodestyle==2.7.0
pycparser==2.20
PyJWT==2.0.1
python-magic==0.4.22
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.4
